Reorganize cats compatibility module
